Kill the core.
- Move /sexp to /ast
- Move ast_node, ast_node_class_map, object_refs and reference_collector to /ast
- Move code_comment, examiner, tree_walker and tree_dresser to top level
- Fully qualify calls to AST::Node via ::Parser to make it obvious that
  we have Reek::AST::ASTNode and Parser::AST::Node
