+++ lib/reek.rb
@@ -2,5 +2,5 @@
 # Reek's core functionality
 #
 require_relative 'reek/version'
-require_relative 'reek/core/examiner'
+require_relative 'reek/examiner'
 require_relative 'reek/smells/smell_warning'
