+++ lib/reek/ast/ast_node.rb
@@ -1,10 +1,10 @@
 require 'parser'
 
 module Reek
-  module Core
+  module AST
     # Base class for AST nodes extended with utility methods. Contains some
     # methods to ease the transition from Sexp to AST::Node.
-    class ASTNode < Parser::AST::Node
+    class ASTNode < ::Parser::AST::Node
       def initialize(type, children = [], options = {})
         @comments = options.fetch(:comments, [])
         super
