+++ lib/reek/ast/sexp_formatter.rb
@@ -4,7 +4,7 @@ Reek::CLI::Silencer.silently do
 end
 
 module Reek
-  module Sexp
+  module AST
     #
     # Formats snippets of syntax tree back into Ruby source code.
     #
@@ -16,7 +16,7 @@ module Reek
       #
       # Returns a formatted string representation.
       def self.format(sexp)
-        return sexp.to_s unless sexp.is_a? AST::Node
+        return sexp.to_s unless sexp.is_a? ::Parser::AST::Node
         lines = Unparser.unparse(sexp).split "\n"
         case lines.length
         when 1 then lines.first
