+++ lib/reek/cli/reek_command.rb
@@ -1,5 +1,5 @@
 require_relative 'command'
-require_relative '../core/examiner'
+require_relative '../examiner'
 
 module Reek
   module CLI
@@ -10,7 +10,7 @@ module Reek
     class ReekCommand < Command
       def execute(app)
         @options.sources.each do |source|
-          reporter.add_examiner Core::Examiner.new(source, smell_names)
+          reporter.add_examiner Examiner.new(source, smell_names)
         end
         reporter.smells? ? app.report_smells : app.report_success
         reporter.show
