+++ lib/reek/context/module_context.rb
@@ -1,5 +1,5 @@
 require_relative 'code_context'
-require_relative '../sexp/sexp_formatter'
+require_relative '../ast/sexp_formatter'
 
 module Reek
   module Context
@@ -9,7 +9,7 @@ module Reek
     class ModuleContext < CodeContext
       def initialize(outer, exp)
         super(outer, exp)
-        @name = Sexp::SexpFormatter.format(exp.children.first)
+        @name = AST::SexpFormatter.format(exp.children.first)
       end
 
       def node_instance_methods
@@ -17,7 +17,7 @@ module Reek
       end
 
       def descriptively_commented?
-        Core::CodeComment.new(exp.leading_comment).descriptive?
+        CodeComment.new(exp.leading_comment).descriptive?
       end
     end
   end
