+++ lib/reek/spec/should_reek_of.rb
@@ -1,4 +1,4 @@
-require_relative '../core/examiner'
+require_relative '../examiner'
 
 module Reek
   module Spec
@@ -13,7 +13,7 @@ module Reek
       end
 
       def matches?(actual)
-        @examiner = Core::Examiner.new(actual)
+        @examiner = Examiner.new(actual)
         @all_smells = @examiner.smells
         @all_smells.any? { |warning| warning.matches?(@smell_category, @smell_details) }
       end
