+++ spec/reek/cli/html_report_spec.rb
@@ -1,12 +1,12 @@
 require_relative '../../spec_helper'
-require_relative '../../../lib/reek/core/examiner'
+require_relative '../../../lib/reek/examiner'
 require_relative '../../../lib/reek/cli/report/report'
 
 RSpec.describe Reek::CLI::Report::HTMLReport do
   let(:instance) { Reek::CLI::Report::HTMLReport.new }
 
   context 'with an empty source' do
-    let(:examiner) { Reek::Core::Examiner.new('') }
+    let(:examiner) { Reek::Examiner.new('') }
 
     before do
       instance.add_examiner examiner
