+++ spec/reek/cli/yaml_report_spec.rb
@@ -1,5 +1,5 @@
 require_relative '../../spec_helper'
-require_relative '../../../lib/reek/core/examiner'
+require_relative '../../../lib/reek/examiner'
 require_relative '../../../lib/reek/cli/report/report'
 require_relative '../../../lib/reek/cli/report/formatter'
 
@@ -7,7 +7,7 @@ RSpec.describe Reek::CLI::Report::YAMLReport do
   let(:instance) { Reek::CLI::Report::YAMLReport.new }
 
   context 'empty source' do
-    let(:examiner) { Reek::Core::Examiner.new('') }
+    let(:examiner) { Reek::Examiner.new('') }
 
     before do
       instance.add_examiner examiner
