+++ spec/reek/smells/duplicate_method_call_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
 require_relative '../../../lib/reek/smells/duplicate_method_call'
 require_relative '../../../lib/reek/context/code_context'
-require_relative '../../../lib/reek/core/tree_walker'
+require_relative '../../../lib/reek/tree_walker'
 require_relative 'smell_detector_shared'
 
 RSpec.describe Reek::Smells::DuplicateMethodCall do
