+++ spec/reek/tree_walker_spec.rb
@@ -1,7 +1,7 @@
-require_relative '../../spec_helper'
-require_relative '../../../lib/reek/core/tree_walker'
+require_relative '../spec_helper'
+require_relative '../../lib/reek/tree_walker'
 
-RSpec.describe Reek::Core::TreeWalker, 'with no method definitions' do
+RSpec.describe Reek::TreeWalker, 'with no method definitions' do
   it 'reports no problems for empty source code' do
     expect('').not_to reek
   end
@@ -11,14 +11,14 @@ class Fred; end').not_to reek
   end
 end
 
-RSpec.describe Reek::Core::TreeWalker, 'with a global method definition' do
+RSpec.describe Reek::TreeWalker, 'with a global method definition' do
   it 'reports no problems for simple method' do
     src = 'def Outermost::fred() true; end'
     expect(src).not_to reek
   end
 end
 
-RSpec.describe Reek::Core::TreeWalker, 'when a yield is the receiver' do
+RSpec.describe Reek::TreeWalker, 'when a yield is the receiver' do
   it 'reports no problems' do
     src = <<EOS
 def values(*args)
