+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require_relative '../lib/reek/spec'
-require_relative '../lib/reek/core/ast_node_class_map'
+require_relative '../lib/reek/ast/ast_node_class_map'
 require_relative '../lib/reek/configuration/app_configuration'
 
 Reek::CLI::Silencer.silently do
@@ -37,7 +37,7 @@ module Helpers
 
   # :reek:UncommunicativeMethodName
   def s(type, *children)
-    @klass_map ||= Reek::Core::ASTNodeClassMap.new
+    @klass_map ||= Reek::AST::ASTNodeClassMap.new
     @klass_map.klass_for(type).new(type, children)
   end
